[Patch] simplified app.py file and fixed a bug
