+++ app.py
@@ -1,19 +1,16 @@
-import json
-import os
 import logging
-import requests
-import openai
 from quart import (
     Blueprint,
     Quart,
     jsonify,
-    Response,
     request,
     send_from_directory
 )
-from dotenv import load_dotenv
-
-load_dotenv()
+from backend.settings import app_settings
+from backend.utils.azure import (
+    conversation_with_data,
+    conversation_without_data
+)
 
 bp = Blueprint("routes", __name__, static_folder="static", template_folder="static")
 
@@ -23,297 +20,32 @@ def create_app():
     app.config["TEMPLATES_AUTO_RELOAD"] = True
     return app
 
-# Static Files
+# Static File Routes
 @bp.route("/")
-def index():
-    return bp.send_static_file("index.html")
+async def index():
+    return await bp.send_static_file("index.html")
 
 @bp.route("/favicon.ico")
-def favicon():
-    return bp.send_static_file('favicon.ico')
+async def favicon():
+    return await bp.send_static_file('favicon.ico')
 
 @bp.route("/assets/<path:path>")
-def assets(path):
-    return send_from_directory("static/assets", path)
+async def assets(path):
+    return await send_from_directory("static/assets", path)
 
-
-# ACS Integration Settings
-AZURE_SEARCH_SERVICE = os.environ.get("AZURE_SEARCH_SERVICE")
-AZURE_SEARCH_INDEX = os.environ.get("AZURE_SEARCH_INDEX")
-AZURE_SEARCH_KEY = os.environ.get("AZURE_SEARCH_KEY")
-AZURE_SEARCH_USE_SEMANTIC_SEARCH = os.environ.get("AZURE_SEARCH_USE_SEMANTIC_SEARCH", "false")
-AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG = os.environ.get("AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG", "default")
-AZURE_SEARCH_TOP_K = os.environ.get("AZURE_SEARCH_TOP_K", 5)
-AZURE_SEARCH_ENABLE_IN_DOMAIN = os.environ.get("AZURE_SEARCH_ENABLE_IN_DOMAIN", "true")
-AZURE_SEARCH_CONTENT_COLUMNS = os.environ.get("AZURE_SEARCH_CONTENT_COLUMNS")
-AZURE_SEARCH_FILENAME_COLUMN = os.environ.get("AZURE_SEARCH_FILENAME_COLUMN")
-AZURE_SEARCH_TITLE_COLUMN = os.environ.get("AZURE_SEARCH_TITLE_COLUMN")
-AZURE_SEARCH_URL_COLUMN = os.environ.get("AZURE_SEARCH_URL_COLUMN")
-AZURE_SEARCH_VECTOR_COLUMNS = os.environ.get("AZURE_SEARCH_VECTOR_COLUMNS")
-AZURE_SEARCH_QUERY_TYPE = os.environ.get("AZURE_SEARCH_QUERY_TYPE")
-AZURE_SEARCH_PERMITTED_GROUPS_COLUMN = os.environ.get("AZURE_SEARCH_PERMITTED_GROUPS_COLUMN")
-
-# AOAI Integration Settings
-AZURE_OPENAI_RESOURCE = os.environ.get("AZURE_OPENAI_RESOURCE")
-AZURE_OPENAI_MODEL = os.environ.get("AZURE_OPENAI_MODEL")
-AZURE_OPENAI_ENDPOINT = os.environ.get("AZURE_OPENAI_ENDPOINT")
-AZURE_OPENAI_KEY = os.environ.get("AZURE_OPENAI_KEY")
-AZURE_OPENAI_TEMPERATURE = os.environ.get("AZURE_OPENAI_TEMPERATURE", 0)
-AZURE_OPENAI_TOP_P = os.environ.get("AZURE_OPENAI_TOP_P", 1.0)
-AZURE_OPENAI_MAX_TOKENS = os.environ.get("AZURE_OPENAI_MAX_TOKENS", 1000)
-AZURE_OPENAI_STOP_SEQUENCE = os.environ.get("AZURE_OPENAI_STOP_SEQUENCE")
-AZURE_OPENAI_SYSTEM_MESSAGE = os.environ.get("AZURE_OPENAI_SYSTEM_MESSAGE", "You are an AI assistant that helps people find information.")
-AZURE_OPENAI_PREVIEW_API_VERSION = os.environ.get("AZURE_OPENAI_PREVIEW_API_VERSION", "2023-06-01-preview")
-AZURE_OPENAI_STREAM = os.environ.get("AZURE_OPENAI_STREAM", "true")
-AZURE_OPENAI_MODEL_NAME = os.environ.get("AZURE_OPENAI_MODEL_NAME", "gpt-35-turbo") # Name of the model, e.g. 'gpt-35-turbo' or 'gpt-4'
-AZURE_OPENAI_EMBEDDING_ENDPOINT = os.environ.get("AZURE_OPENAI_EMBEDDING_ENDPOINT")
-AZURE_OPENAI_EMBEDDING_KEY = os.environ.get("AZURE_OPENAI_EMBEDDING_KEY")
-
-
-SHOULD_STREAM = True if AZURE_OPENAI_STREAM.lower() == "true" else False
-
-def is_chat_model():
-    if 'gpt-4' in AZURE_OPENAI_MODEL_NAME.lower() or AZURE_OPENAI_MODEL_NAME.lower() in ['gpt-35-turbo-4k', 'gpt-35-turbo-16k']:
-        return True
-    return False
-
-def format_as_ndjson(obj: dict) -> str:
-    return json.dumps(obj, ensure_ascii=False) + "\n"
-
-def fetchUserGroups(userToken, nextLink=None):
-    # Recursively fetch group membership
-    if nextLink:
-        endpoint = nextLink
-    else:
-        endpoint = "https://graph.microsoft.com/v1.0/me/transitiveMemberOf?$select=id"
-    
-    headers = {
-        'Authorization': "bearer " + userToken
-    }
-    try :
-        r = requests.get(endpoint, headers=headers)
-        if r.status_code != 200:
-            return []
-        
-        r = r.json()
-        if "@odata.nextLink" in r:
-            nextLinkData = fetchUserGroups(userToken, r["@odata.nextLink"])
-            r['value'].extend(nextLinkData)
-        
-        return r['value']
-    except Exception as e:
-        return []
-
-
-def generateFilterString(userToken):
-    # Get list of groups user is a member of
-    userGroups = fetchUserGroups(userToken)
-
-    # Construct filter string
-    if userGroups:
-        group_ids = ", ".join([obj['id'] for obj in userGroups])
-        return f"{AZURE_SEARCH_PERMITTED_GROUPS_COLUMN}/any(g:search.in(g, '{group_ids}'))"
-    
-    return None
-
-
-def prepare_body_headers_with_data(request):
-    request_messages = request.get("messages", [])
-
-    # Set query type
-    query_type = "simple"
-    if AZURE_SEARCH_QUERY_TYPE:
-        query_type = AZURE_SEARCH_QUERY_TYPE
-    elif AZURE_SEARCH_USE_SEMANTIC_SEARCH.lower() == "true" and AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG:
-        query_type = "semantic"
-
-    # Set filter
-    filter = None
-    userToken = None
-    if AZURE_SEARCH_PERMITTED_GROUPS_COLUMN:
-        userToken = request.headers.get('X-MS-TOKEN-AAD-ACCESS-TOKEN', "")
-        filter = generateFilterString(userToken)
-
-    body = {
-        "messages": request_messages,
-        "temperature": float(AZURE_OPENAI_TEMPERATURE),
-        "max_tokens": int(AZURE_OPENAI_MAX_TOKENS),
-        "top_p": float(AZURE_OPENAI_TOP_P),
-        "stop": AZURE_OPENAI_STOP_SEQUENCE.split("|") if AZURE_OPENAI_STOP_SEQUENCE else None,
-        "stream": SHOULD_STREAM,
-        "dataSources": [
-            {
-                "type": "AzureCognitiveSearch",
-                "parameters": {
-                    "endpoint": f"https://{AZURE_SEARCH_SERVICE}.search.windows.net",
-                    "key": AZURE_SEARCH_KEY,
-                    "indexName": AZURE_SEARCH_INDEX,
-                    "fieldsMapping": {
-                        "contentFields": AZURE_SEARCH_CONTENT_COLUMNS.split("|") if AZURE_SEARCH_CONTENT_COLUMNS else [],
-                        "titleField": AZURE_SEARCH_TITLE_COLUMN if AZURE_SEARCH_TITLE_COLUMN else None,
-                        "urlField": AZURE_SEARCH_URL_COLUMN if AZURE_SEARCH_URL_COLUMN else None,
-                        "filepathField": AZURE_SEARCH_FILENAME_COLUMN if AZURE_SEARCH_FILENAME_COLUMN else None,
-                        "vectorFields": AZURE_SEARCH_VECTOR_COLUMNS.split("|") if AZURE_SEARCH_VECTOR_COLUMNS else []
-                    },
-                    "inScope": True if AZURE_SEARCH_ENABLE_IN_DOMAIN.lower() == "true" else False,
-                    "topNDocuments": AZURE_SEARCH_TOP_K,
-                    "queryType": query_type,
-                    "semanticConfiguration": AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG if AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG else "",
-                    "roleInformation": AZURE_OPENAI_SYSTEM_MESSAGE,
-                    "embeddingEndpoint": AZURE_OPENAI_EMBEDDING_ENDPOINT,
-                    "embeddingKey": AZURE_OPENAI_EMBEDDING_KEY,
-                    "filter": filter
-                }
-            }
-        ]
-    }
-
-    headers = {
-        'Content-Type': 'application/json',
-        'api-key': AZURE_OPENAI_KEY,
-        "x-ms-useragent": "GitHubSampleWebApp/PublicAPI/2.0.0"
-    }
-
-    return body, headers
-
-
-def stream_with_data(body, headers, endpoint):
-    s = requests.Session()
-    response = {
-        "id": "",
-        "model": "",
-        "created": 0,
-        "object": "",
-        "choices": [{
-            "messages": []
-        }]
-    }
-    try:
-        with s.post(endpoint, json=body, headers=headers, stream=True) as r:
-            for line in r.iter_lines(chunk_size=10):
-                if line:
-                    lineJson = json.loads(line.lstrip(b'data:').decode('utf-8'))
-                    if 'error' in lineJson:
-                        yield format_as_ndjson(lineJson)
-                    response["id"] = lineJson["id"]
-                    response["model"] = lineJson["model"]
-                    response["created"] = lineJson["created"]
-                    response["object"] = lineJson["object"]
-
-                    role = lineJson["choices"][0]["messages"][0]["delta"].get("role")
-                    if role == "tool":
-                        response["choices"][0]["messages"].append(lineJson["choices"][0]["messages"][0]["delta"])
-                    elif role == "assistant": 
-                        response["choices"][0]["messages"].append({
-                            "role": "assistant",
-                            "content": ""
-                        })
-                    else:
-                        deltaText = lineJson["choices"][0]["messages"][0]["delta"]["content"]
-                        if deltaText != "[DONE]":
-                            response["choices"][0]["messages"][1]["content"] += deltaText
-
-                    yield format_as_ndjson(response)
-    except Exception as e:
-        yield format_as_ndjson({"error": str(e)})
-
-
-def conversation_with_data(request):
-    body, headers = prepare_body_headers_with_data(request)
-    base_url = AZURE_OPENAI_ENDPOINT if AZURE_OPENAI_ENDPOINT else f"https://{AZURE_OPENAI_RESOURCE}.openai.azure.com/"
-    endpoint = f"{base_url}openai/deployments/{AZURE_OPENAI_MODEL}/extensions/chat/completions?api-version={AZURE_OPENAI_PREVIEW_API_VERSION}"
-    
-    if not SHOULD_STREAM:
-        r = requests.post(endpoint, headers=headers, json=body)
-        status_code = r.status_code
-        r = r.json()
-
-        return Response(format_as_ndjson(r), status=status_code)
-    else:
-        return Response(stream_with_data(body, headers, endpoint))
-
-def stream_without_data(response):
-    responseText = ""
-    for line in response:
-        deltaText = line["choices"][0]["delta"].get('content')
-        if deltaText and deltaText != "[DONE]":
-            responseText += deltaText
-
-        response_obj = {
-            "id": line["id"],
-            "model": line["model"],
-            "created": line["created"],
-            "object": line["object"],
-            "choices": [{
-                "messages": [{
-                    "role": "assistant",
-                    "content": responseText
-                }]
-            }]
-        }
-        yield format_as_ndjson(response_obj)
-
-
-def conversation_without_data(request):
-    openai.api_type = "azure"
-    openai.api_base = AZURE_OPENAI_ENDPOINT if AZURE_OPENAI_ENDPOINT else f"https://{AZURE_OPENAI_RESOURCE}.openai.azure.com/"
-    openai.api_version = "2023-03-15-preview"
-    openai.api_key = AZURE_OPENAI_KEY
-
-    request_messages = request.get("messages", [])
-
-    messages = [
-        {
-            "role": "system",
-            "content": AZURE_OPENAI_SYSTEM_MESSAGE
-        }
-    ]
-
-    for message in request_messages:
-        messages.append({
-            "role": message["role"] ,
-            "content": message["content"]
-        })
-
-    response = openai.ChatCompletion.create(
-        engine=AZURE_OPENAI_MODEL,
-        messages = messages,
-        temperature=float(AZURE_OPENAI_TEMPERATURE),
-        max_tokens=int(AZURE_OPENAI_MAX_TOKENS),
-        top_p=float(AZURE_OPENAI_TOP_P),
-        stop=AZURE_OPENAI_STOP_SEQUENCE.split("|") if AZURE_OPENAI_STOP_SEQUENCE else None,
-        stream=SHOULD_STREAM
-    )
-
-    if not SHOULD_STREAM:
-        response_obj = {
-            "id": response,
-            "model": response.model,
-            "created": response.created,
-            "object": response.object,
-            "choices": [{
-                "messages": [{
-                    "role": "assistant",
-                    "content": response.choices[0].message.content
-                }]
-            }]
-        }
-
-        return jsonify(response_obj), 200
-    else:
-        return Response(stream_without_data(response))
-
+# App Routes
 @bp.route("/conversation", methods=["GET", "POST"])
-def conversation():
+async def conversation():
     try:
-        if AZURE_SEARCH_SERVICE and AZURE_SEARCH_INDEX and AZURE_SEARCH_KEY:
-            result = conversation_with_data(request)
+        if app_settings.search.service and app_settings.search.index and app_settings.search.key:
+            result = await conversation_with_data(request)
             print(str(result))
             return result
         else:
-            return conversation_without_data(request)
+            return await conversation_without_data(request)
     except Exception as e:
         logging.exception("Exception in /conversation")
         return jsonify({"error": str(e)}), 500
 
+# Init App
 app = create_app()
